Use simple string for memory blob
Mandating use of SimpleJSON is not beneficial because other methods for
JSON handling (e.g.
https://docs.unity3d.com/Manual/JSONSerialization.html) may be used
instead.

It might be possible to automate handling with `JsonUtility`, but again
it depends on which kind of data is serialized on whether or not this is
suitable. Just passing around JSON as strings seems most compatible with
different preferred styles.