Fixed issues with WebGL.
Issues were related to WebGL not being able to infer types across assemblies. The DividatSenso package is compiled into a different assembly than the project, and so the types weren't visible. As long as the savegame type is a subclass of PlaySaveGame AND is part of the project code, and not the DividatSenso Plugin package, this setup works well and allows to completely abstract the the memory class from the JSON.